+++ src/base/b-router/interface.ts
@@ -32,7 +32,7 @@ export {
 
 } from 'core/router/interface';
 
-export type { LinkNavigateEvent } from 'base/b-router/modules/transition/event';
+export type { HrefTransitionEvent } from 'base/b-router/modules/transition/event';
 
 export type AppliedRoute = EngineAppliedRoute<
 	bRouter['PageParams'],